# Simplify intrinsic trimming